+++ docs/source/changelog.md
@@ -7,6 +7,8 @@
 
 - Removed BETA status from the site.
     - Also changed the welcome email templates to remove the BETA status.
+- Upload form
+    - Improved validation of the AdditionalTitlesField and AdditionalDescriptionsField.
 - Bug fixes
     - Improved the search used by the "My Works" dashboard tab.
     - Fixed bugs in upload form error handling.
@@ -14,6 +16,8 @@
     - Fixed a bug where the languages field was not being populated correctly from the client-side form state.
     - Updated the language subfield of AdditionalTitlesField and AdditionalDescriptionsField to use a new SingleLanguageSelector component
       that also populates and preserves the value properly from the client-side form state.
+- Themeing
+    - Minor tweaks to the alignment in AdditionalDescriptionsField
 - Testing
     - Added new jest tests for upload form validation and the languages field.